Bugfixes (#7)
* Replaced the change event with input

'change' gets only called when the input is finished, i.e. if the element loses its focus. With 'change' the event gets called on every button press.

* Importing the same file a second time works

It was not possible to import the same file again, after you already imported it. Now you can import an file, clear the work space and re-import it again.

* Added a safety popup before deleting a dialog

Added a pop up if you use the "new"-button to confirm that you want to clear the current dialog (and might lose the data) and create a new dialog.

* Fixed the import type

The dialogs are exported as *.json, but the import file explorer was looking for *.dl files.